Eliminate the duplicate connection name message
This patch eliminates the "QSqlDatabasePrivate::addDatabase: duplicate
connection name 'qt_sql_default_connection', old connection removed."
warning message.

This patch replaces the Database implementation with a Database class
that permits the database to be open and closed multiple times by the
application.

This class implementation effectively moves init() to the constructor,
and init() to the execute() method.  The populate() method is effectively
replaced by execute(). A class is needed to manage the connection name.