Added database argument to Schema's migrate method
The Schema class's migrate() method must take a database argument to
enable database migration. An argument to the migrate() method imposes
two conditions. First client's using the TableRevisionHistroy class must
pass in the database that the history will be executed against. Correct
use of the class assumes the use of this class to achieve that result.

Using TableRevision could achieve a similar result but isn't the correct
way to use the Schema.